Move CLI tests to corpus as integration tests
