--- cli/cli_bal_run_test.go
@@ -1,105 +1,0 @@
-// Copyright (c) 2026, WSO2 LLC. (http://www.wso2.com).
-//
-// WSO2 LLC. licenses this file to you under the Apache License,
-// Version 2.0 (the "License"); you may not use this file except
-// in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package main
-
-import (
-	"flag"
-	"os"
-	"path/filepath"
-	"runtime"
-	"sort"
-	"strconv"
-	"strings"
-	"testing"
-)
-
-func TestBalRunCorpus(t *testing.T) {
-	if runtime.GOOS == "js" || runtime.GOARCH == "wasm" {
-		t.Skip("skipping CLI integration test on WASM (js/wasm)")
-	}
-	flag.Parse()
-
-	balBin, repoRoot, coverDir := integrationTestBal(t, false)
-	testDataRoot := filepath.Join(repoRoot, "corpus", "cli", "testdata", "run")
-	outputsRoot := filepath.Join(repoRoot, "corpus", "cli", "output", "run")
-
-	singleBalFiles := listPaths(t, filepath.Join(testDataRoot, "single-bal-files"), true)
-	projects := listPaths(t, filepath.Join(testDataRoot, "projects"), false)
-
-	for _, singleBalFile := range singleBalFiles {
-		rel := filepath.Join("single-bal-files", strings.TrimSuffix(filepath.Base(singleBalFile), ".bal"))
-		runAndValidateCase(t, balBin, repoRoot, coverDir, outputsRoot, singleBalFile, rel)
-	}
-
-	for _, projectDir := range projects {
-		rel := filepath.Join("projects", filepath.Base(projectDir))
-		runAndValidateCase(t, balBin, repoRoot, coverDir, outputsRoot, projectDir, rel)
-	}
-}
-
-func listPaths(t *testing.T, dir string, balFilesOnly bool) []string {
-	t.Helper()
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		t.Fatalf("failed to read directory %s: %v", dir, err)
-	}
-	paths := make([]string, 0)
-	for _, entry := range entries {
-		if balFilesOnly {
-			if entry.IsDir() || !strings.HasSuffix(entry.Name(), ".bal") {
-				continue
-			}
-		} else if !entry.IsDir() {
-			continue
-		}
-		paths = append(paths, filepath.Join(dir, entry.Name()))
-	}
-	sort.Strings(paths)
-	return paths
-}
-
-func runAndValidateCase(t *testing.T, balBin, repoRoot, coverDir, outputsRoot, runPath, outputKey string) {
-	t.Helper()
-	t.Run(strings.ReplaceAll(outputKey, string(filepath.Separator), "_"), func(t *testing.T) {
-		stdout, stderr, exitCode := runCLICommand(t, balBin, repoRoot, coverDir, "run", runPath)
-		expectedPath := filepath.Join(outputsRoot, outputKey+".txtar")
-		actualOutput := normalizeNewlines(stdout)
-		actualError := normalizeNewlines(stderr)
-		actualExitCode := strconv.Itoa(exitCode)
-
-		if *update {
-			if updateOutputArchiveIfNeeded(t, expectedPath, actualOutput, actualError, actualExitCode) {
-				t.Fatalf("Updated expected file: %s", expectedPath)
-			}
-			return
-		}
-
-		expectedOutput, expectedError, expectedExitCode := readExpectedTxtar(t, expectedPath)
-		if expectedOutput != actualOutput || expectedError != actualError || expectedExitCode != actualExitCode {
-			t.Fatalf(
-				"unexpected output for %s\nexpected stdout:\n%s\nactual stdout:\n%s\nexpected stderr:\n%s\nactual stderr:\n%s\nexpected exitcode: %s\nactual exitcode: %s",
-				runPath,
-				expectedOutput,
-				actualOutput,
-				expectedError,
-				actualError,
-				expectedExitCode,
-				actualExitCode,
-			)
-		}
-	})
-}
